langchain.py学习/requirements.txt
langchain
langchain-ollama
python-dotenv
jupyter
ipykernel
